Issue 3379: move all tablature related settings to TabStaff
Previous to this patch, settings relevant to tablature have been
distributed among TabVoice and TabVoice, presumably modeled along the
corresponding distribution of engravers.

As one consequence, the command \tabFullNotation contained overrides
pertaining to both TabStaff and TabVoice level, leading to
inconsistent behavior when employed in one but not all TabVoice
contexts of a TabStaff, or when used as
\new TabStaff \with { \tabFullNotation }

Moving all overrides and property settings to TabStaff level should
lead to more consistent results.