Issue 3449: Shift accordion register symbols down by half a line width
This makes the documented proposals for combining the accordion
register symbols together with the accordion dot symbol match in
positioning.  It would have been feasible instead to shift the
(centered) accordion dot, but some register symbols work with
different line widths.

The net result will now exhibit a slight descender (namely half a
linewidth) of the accordion register symbols when used in a running
line.