Issue 3488: lily/parser.yy: Use %pure-parser instead of %pure_parser
While %pure-parser is already deprecated for Bison 3.0, it is still
supported as opposed to %pure_parser which has been removed altogether
in Bison 3.0.  The currently recommended form %define api.pure has
only become available with Bison 2.3b, and there are currently older
versions of Bison in use with LilyPond.  So the followup move of
replacing the already deprecated %pure-parser with %define api.pure is
left for a later point of time.