Issue 3648/1: Isolated durations in music sequences now stand for unpitched notes
This allows for things like

\new RhythmicStaff { 4 4. 8 4 | 4 \tuplet 3/2 { 2 4 } 4 }

"unpitched" simply means that this is rendered as a normal NoteEvent,
but the pitch information is missing and has to be filled in by some
other means.  The example using RhythmicStaff is not actually
commendable as the MIDI rendition still requires a pitch or drum
type.  It's just that as of this patch, there are no immediately
useful applications.  The rhythmic shorthand { c4 4. 8 4 } is
introduced by a later patch.