Issue 3554: Avoid empty grammar rules using @$ for getting a source location
Several rules use @$ for getting the source position from an empty
rule, but Bison does not even provide a source position for empty
rules.

This patch fixes most cases by letting different non-empty rules
assign the respective source positions.  The case of #{ #} cannot be
fixed in that manner since there is not a single token providing a
source position in the whole production.  Fixing this particular case
would require a different technique.