Doc: Relayout the download pages (issue 3595).
This slightly modifies the layout of the download pages to make
them more balanced (in the HTML version) and to accomodate wide
pictures.

For Unix, the "Specific packages" box is moved to the top right,
next to the "Generic packages" box, while the "Compiling a file"
box is now a centered, full-width box.

For MacOS and Windows, a similar change is made: the "Compiling a
file" box is moved below and turned into a centered, full-width box
to better accomodate the picture-full content.