Revert "Issue 3205: opening bar check causes crash if \score contains the \midi block"
This reverts commit c83a3afbc51c62769cbb8963371d37b312bbcd42.

The replacement of Simple_music_iterator with Music_iterator in
lily/bar-check-iterator.cc made bar checks at the start of music
expressions inoperative altogether, even when not at a bar boundary:

{ a a { | } a a }

stopped producing a diagnostic.  Issue 3205 instead needs to be fixed
together with issue 2392 which apparently may be called "Segfault in
Midi when Simple_music_iterator is used at the beginning of an
implicitly created context".