Issue 3644: A \score-lines markup list command for multi-lines embedded scores
Originally requested as issue 1334.

Like the \score markup, \score-lines is not called through the normal
markup list command mechanisms as it would require something awkward
along the lines of

\score-lines ##{ \score { ... } #}

to get this through.  Instead, a reserved word \score-lines in the
parser will manually call the command with an embedded score, leading
to just

\score-lines { ... }

This is somewhat clumsy in several ways: it requires an additional
reserved identifier (no other reserved identifier contains a dash),
and it means that a score may occur in a file without an explicit
enclosing \score command.

Nevertheless, the previous semantics implemented as issue 1334 were
not tenable: particularly after issue 3270 had been passed, it was
close to impossible for the average user to come up with a way of
getting an actual markup list from a score.

Note that there are as of yet no user-level commands spacing a markup
list in a way similar to the spacing inside of a \score markup.