Issue 3663: Crash with \repeat ... \alternative and \remove "Bar_engraver"
This approach is just poking around in the dark and removing Scheme
error messages.  In particular substituting the original glyph for
left-bar-line and right-bar-line (when missing) in calls to
span-bar::compound-bar-line seems fishy, and the original logic
for setting left-bar-broken if left-bar-line is not even present
is also not clear to me.

The results for the example in the bug report:
form = \new Staff \with {
  \remove "Bar_engraver"
} \repeat volta 2 {
    s1
} \alternative {
    s1 % first ending
    s1 % second ending
} % form

\score {
  <<
    \form
  >>
} % score
are also not identical with the 2.16 results: the volta brackets now
run into each other; before there was a gap.

It fixes the crash.  I have no idea what a proper fix would look like.