Grace notes: only shorten previous note if overlapping
Previously a grace note after a rest shortened the last note before the
rest.

This patch shortens the previous note to the start of the grace note
rather than by the length of the grace note, so notes before rests
aren't affected unless the grace notes are longer than the rest.