let RhythmicStaff use default barline extent
There is no reason to make barlines in RhythmicStaff special.
The purpose of the deleted code was probably to ensure that
the barlines were visible at all (despite the fact that there
is just one staffline in a RhythmicStaff).  However, currently
barline code takes care of this by default.

This makes barlines in DrumStaff and RhythmicStaff similar.

A comment by Keith:
The book by Kurt Stone happens to have some bar-lines in single-line percussion
staves that are as tall as a 4/4 time-signature.  Other examples in the same
book have short bar-lines like these here, and looking at some scores the short
bar-lines are more common.