note-spacing: let compressibility be uniform; issue 3304
Comments implied that half the width of accidentals, etc., was added
to the ideal spacing between notes, but in fact only compressibility
was affected.  The non-uniform compressibility caused note-spacing
to become non-uniform when the lines were compressed.  For example
A sequence {\stemUp a a d a] would have the head of the D tuck under
the preceding A.

This commit keeps spacing uniform on compressed lines until objects
come within padding of each other.