Improve positioning of tuplet numbers for kneed beams.
In the past LilyPond always placed tuplet numbers according to the
position of a bracket, though this bracket may or may not have actually
been drawn.  This system led at times to tuplet numbers which appeared
stranded when there was a kneed beam and no visible bracket.

This patch positions the tuplet number independently of brackets and
close to the beam when the the bracket is not drawn, the beam is kneed, and
there is an appropriate beam segment to place the number against.

The side of the beam on which the number is placed is chosen by principles
derived from Gould.  The number is centered horizontally on the beam.

Collision detection is also introduced.  The number will be offset
horizontally if it is too close to an adjoining note column.  All shifts
preserve the number's vertical distance from the beam.  If the number is
too large to fit in the available horizontal space (between note columns
to either side), we use the old bracket-based positioning system.  In the
event of a collision with an accidental, the tuplet number is moved
vertically, away from the beam.

The older behavior is always available through an override of the property
`knee-to-beam' which is introduced by this patch, and set by default to true.

This patch also adds a header file, `tuplet-number.hh', and provides
a number of regression tests illustrating various features.