Issue 3785:Web:Background: Reword intr. paragraph
I've always found this page confusing because I didn't understand
why there is this "placeholder" page.
This commit addresses this by making the "manual" issue explicit.

Additionally I consider it more useful to _first_ tell the reader
what it is and _then_ advise him to skip the essay instead of
_immediately_ telling him "Probably you won't want to read this".
If it is necessary to have the link at the end of a paragraph
(and not only of a sentence) I'd simply insert a new paragraph
after the link.