Issue 3815: Parser: harmonize \lyricsto and \addlyrics arguments
Previously, they allowed arbitrary music but since that required looking
at lookahead tokens in the wrong mode, it led to strange errors like with
lyr=\lyricsto A { }
mus={c}
when mus= was parsed as a single token in lyrics mode.  Now both \addlyrics
and \lyricsto accept merely the same kind of delimited argument list
that \lyrics or \chords accept.  Additionally, to preserve compatibility
to a lot of examples, music identifiers like \mus are permitted as the
argument to \lyricsto and \addlyrics.

Since this requires a change in existing uses, this includes a
convert-ly rule that removes redundant uses of \lyricmode and
rearranges the combination with context starters such that \lyricsto
in general is applied last, like \lyricmode would be.