Issue 3826: Stop scripts/auxiliar/update-with-convert-ly.sh from doing a full make
It is quite a nuisance for writing convert-ly rules that
scripts/auxiliar/update-with-convert-ly runs a full make since any
partly successful run of update-with-convert-ly will cause large
delays.  It also leads to a paradoxical situation when the syntax of
some files needs to be converted _before_ a full make has a chance to
succeed.

So this patch reduces the preliminary make to "make pythonmodules"
which should be sufficient for updating from convertrules.py changes.