Issue 3860: Add command \justify-line
The command \fill-line spaces words taking word length into account
such that the following distances are equal: outside edge of outside
markup to center of adjoining markup; center to center of
neighboring interior markups.  As a consequence, the space between
words is unequal when the lengths of the markups vary and/or there
are more than three markups in the line.

The following patch creates the command \justify-line, which spaces
markups to fill a line such that the space in between each pair is
constant.  Collisions are prevented by ensuring that this distance
does not fall below the current setting of word-space.