Issue 3876: Fix shape of G clefs
This patch simplifies the code for the double G clefs; it also fixes the
incorrect test of the `double_shift' parameter that caused the non-double
G clefs being drawn twice.

Recent FontForge warnings like

  Internal Error (overlap) in clefs.G: Intersection failed!

or the ugly shape with a straight line instead of a curve in the left part
of the G clef are now gone: Drawing the shape twice triggered a bunch of
FontForge bugs.