Issue 3880: bad number formatting in svg path output
This uses ly:format with "~4f" format string rather than format with
"~S" for formatting point lists.  Note that "~4f" has a different
meaning in ly:format as compared to plain format.  This is consistent
with other number formatting.  Also does this for output of color.

The previous output was problematic when the number was an exact
fraction.