Issue 4487/3: Implement partial markups
A partial markup acts as a chain of markup commands where everything but
the final markup has already been supplied.

For example:

bold-red-markup = \markup \bold \with-color #red \etc

\markup \bold-red "text"
\markuplist \column-lines \bold-red { One Two Three }

Please note that in order for \markup to recognize the resulting markup
command inside of markup expressions, the identifier to which the
resulting markup command is assigned needs to have "-markup" appended to
its name: LilyPond encodes markup command names in this manner so that
markup commands can use the same name as other entities outside of
markup mode.

Note also that this will not work with the markup macro since the latter
additionally requires the definition of a suitable make-bold-red-markup
command.