Doc: NR Section 4.3 - reformatting as per CG Guidelines
Start of issue 3960

This section does not adhere to  many aspects of the Contributors'
Guide Documentation policies.

This specific patch therefore is the 'first stage' of this issue
so as to clean up the section before then making changes to
the technical information. There is too much to do for a single patch,
so I expect to have to make a few more additional patches to tidy this
section up completely before the issue can be properly closed.

This patch includes:

* Correct Line lengths

* Added @seealsos where appropriate

* Re-ordered some of the @nodes so that they are grouped more logically

* Moved @subsections to @unnumberedsubsubsecs:
A few of these sections seem to 'talk through the code' so may
need to be improved later on. However, for now I have moved these
subsections to unnumberedsubsubsec under the main 'Page Breaking'
node. Moved one of these new unnumberedsubsucsecs around so that
the sections on the different breaks come before the turning one.

* Created new Snippet:
One subsection was really a glorified snippet so I have made it into
one.