Doc: NR Pitches.itely - Clef - updated snippets and text
Issue 3976

Added 2 new snippets:
1 snippet added to NR
1 snippet added just to Snippets

Modified 1 existing snippet:
This was already (and still is) included in the NR

Deleted 2 snippets (and incorporated them in the
main text in @lilyponds)

Added new @lilypond example that was originally
described (but not show)n in the 'Tweaking Clef
Properties' snippet to show how to remove
the clef glyph from the end of a line. As this
'tweak' is relatively simple (a single \set and
\unset) and more than likely to be used (IMHO) by
the general user than the listed items in the
snippet, I thought it better to be in the NR than
another example in either an already long snippet
or by creating yet another snippet.

Some minor linelength formatting as per CG.