Doc: NR 4.5.4 Change Node name from 'Line length' to 'Line width'
Issue 3982

Change Node name from 'Line length' to 'Line width'.

In the Notation Reference, this page is called
"Line length" but the property that it covers
(among others) is called "line-width."  Seems like
it would be better if the title of the page was
renamed "Line width" for greater consistency.