Issue 3986: \displayScheme broken
    \displayScheme \markup \null

failed because a markup following an omitted optional argument (here the
optional output port) was backed up erroneously with the token type
LYRIC_ELEMENT only to be used in lyrics mode.

The optional argument was introduced with issue 2067 in version 2.19.0.
The actual underlying parser error triggered by this, however, is quite
older and originates from

    commit d4694123961690b173377b3c3a199d5a24c66a9d
    Author: David Kastrup <dak@gnu.org>
    Date:   Thu Nov 1 15:16:52 2012 +0100

	parser/lexer: eliminate LYRIC_MARKUP as a token separate from MARKUP (2996/2)

in version 2.17.9.