Issue 3518: Support temporary divisi staves
This provides the low-level support for temporary divisi staves by
adding a `VerticalAxisGroup.remove-layer' property of type integer that
interacts with the "Keep_alive_together_engraver": when set to a numeric
value, staves with the same numeric value are kept alive together as one
group.  Of several such groups with live staves, only the one with the
lowest common numeric `remove-layer' is retained.