Issue 4102: misaligned composer in LP 2.19.14
This problem surfaced in the wake of the fix for issue 3855.  It is due
to Stencil::translate not moving an empty expression even if it has
non-empty extents so that

\markup fill-line { "" "" "composer" }

which assembles a line piece by piece and determining the offset of the
next piece by looking at the end of the assemblage so far lost the
offset between the first and second "" in the line.