Issue 4125: Disable character set recognition
GUILE 1.8 presumably does not make use of the LC_CTYPE information and
just does its input/output/string processing in 8bit chunks.  For the
sake of getting comparable results at first when working on GUILE 2.0,
we switch off character set recognition here.

At a later point of time, integration with the character processing of
GUILE 2.0 may be attempted but since this is an entirely new can of
worms with its own performance pitfalls, we put this aside for now.