Roman numerals may be used for page numbers
This patch builds on the number-format function used for
string numbers and enables it for page numbers, through
a new 'page-number-type paper property. Previously,
a custom markup function had to be used as demonstrated
in the bookparts.ly regtest, which has now been rewritten
to make use of the new property.