Improvements to measure counter
This patch adds no new functionality to the measure counter, but
there are improvements to the code.

The stencil callback has been shortened, as there is no need to
search for columns which are already available as the bounds of
broken spanners.  It has also been moved to scm/output-lib.scm,
where other such functions are.

The engraver has been streamlined by the removal of an unnecessary
variable and levels of nesting.  Comments have been revised.