Web: Easier Editing - add LilyBin
Issue 4199

Added LilyBin to the page, also took the time for some
checks and re-organizing of exiting entries.

* Changed the structure slighly so that the sectioning
is now grouped by 'type' of editor rather than ordered
by the name. This allowed the browser-based ones to no
longer need the Operating System icons next to them.

* Removed the embedded URLs to Encore (GNU Policy for
links to non-free software).

* Checked the 'Other Programs' Section to see if the
URLs still worked (fixed those that didn't) and moved
a couple of the entries to the 'no longer being developed
section' based on the 'last updated' dates on their
respective websites.