Issue 4233: Improve partcombine multi-measure rest handling.
Add "silence1" and "silence2" states which constrain the partcombine
iterator to take events from a specific part during intervals of
silence.  These new states allow the following improvements.

Detect when a multi-measure rest starts in the middle of a
multi-measure rest in the other voice and engrave it as a shared rest.
Previously, the new rest would have been engraved as a voice-specific
rest.

In solo analysis, when a rest and a multi-measure rest begin
simultaneously, engrave the rest as a shared rest and omit the
multi-measure rest.  Previously, both rests would have been engraved
as voice-specific rests.