Articulate.ly: Fix mordent and prall timing
Issue 4247

When a prall or mordent was on a note
shorter than a semiquaver, the calculation
of timing for the main note after stealing
the time for the ornament used to generate
a negative duration.

Instead of a subtracting ornament time
from the main note, generate the ornament
using a fixed timing, tack it on the
beginning, and then compress the resulting
music to be the same length as the original.