Doc: CG added notes about Google Account Settings
Issue 4250

Both git-cl and pathcy scripts assume that
the user has a Google account set up.

It seems that the [default] Account setting
'Access for less secure apps' must be set to
'Allowed' for these two tools to work as
expected.

Added an @warning in the two sections that state
the Google account requirement.