Add Scheme function to return column associated with Item.
The C++ function Paper_column::get_column is used frequently,
and it makes sense to have a Scheme function which performs
the same task.  An example of a practical application
is the request by users to return the location of a particular
grob in a score: this is done by examining the PaperColumn
or NonMusicalPaperColumn associated with the grob.