Issue 4294: Functions to display rhythmic information of a grob
Convenient ways to return the musical position of a grob within
a score will be very helpful in debugging, among other uses.

This patch creates two Scheme functions which return the location
of a grob based on the associated paper column.  In the case of
spanners, this is the column of the left bound.

(1) grob::rhythmic-location returns a pair consisting of measure-number
and measure-position;

(2) grob::when returns a moment representing the global timestep.