Issue 4324: Don't create Bottom to announce TimeSignatureEvent
As a fallout of issue 4138, << \time 3/4 \new Staff c2. >> created a
spurious Staff only containing a time signature.  This change avoids the
creation of a Bottom context just for accommodating \time.  As a result,

    { \tweak color #red \time 3/4 c2. }

will no longer show a change in color since no Staff context exists at
the time \time is executed and consequently the tweak does not apply
anywhere.  You have to use \time in existing Staff contexts to have
either \tweak and point&click have an effect (and get an error locator
when something with typesetting the signature goes wrong).  Outside of
such contexts, \time just changes properties of the Timing context.